Initial attempt at recursive splitting.
This is supposed to recursively split the given list and pass the left
portion of the list to be recursively split into smaller elements. What
remains to be done is to merge the split lists. Then once that is
accomplished, the same needs to be done for the right side of the lists.